GEODE-6425: Fixes rollover in some cache stats (#3214)
Changes the following CachePerfStats to long:
 - gets
 - misses
 - put
 - update
 - create
 - destroys
 - invalidates 
 - clear

Handles the error caused by attempting to retrieve a long type stat as an
int. Returns the long value as an int with overflow. This is consistent
with previous behavior of the stats

+++ geode-core/src/distributedTest/java/org/apache/geode/cache30/DistributedNoAckRegionCCEDUnitTest.java
@@ -334,13 +334,13 @@ public class DistributedNoAckRegionCCEDUnitTest extends DistributedNoAckRegionDU
           public void run() {
             CCRegion.put(invalidationKey, "initialValue");
 
-            int invalidationCount = CCRegion.getCachePerfStats().getInvalidates();
+            long invalidationCount = CCRegion.getCachePerfStats().getInvalidates();
             CCRegion.invalidate(invalidationKey);
             CCRegion.invalidate(invalidationKey);
             assertEquals(invalidationCount + 1, CCRegion.getCachePerfStats().getInvalidates());
 
             // also test destroy() while we're at it. It should throw an exception
-            int destroyCount = CCRegion.getCachePerfStats().getDestroys();
+            long destroyCount = CCRegion.getCachePerfStats().getDestroys();
             CCRegion.destroy(invalidationKey);
             try {
               CCRegion.destroy(invalidationKey);
